requirements.txt
streamlit
apify-haystack==0.1.4
haystack-ai==2.10.3
python-dotenv==1.0.1
